Fix the linking JS when apaxy is at the root
The code previously did not gracefully handle the situation when apaxy
was not in a subdirectory, but instead at the root of a domain. In that
situation, it would actually add 2 forward slashes to the URL, causing
the link to become a protocol-relative absolute URL. (e.g. 'example.c'
=> '//example.c')

Easy fix: Skip when the substring is empty.